🌞 Rename subpackage load to loaders (#586)
* refactor: rename load to loaders

* refactor; fix monky patching
