+++ cognite_toolkit/_cdf.py
@@ -26,7 +26,7 @@ from cognite_toolkit._cdf_tk.exceptions import (
     ToolkitInvalidSettingsError,
     ToolkitValidationError,
 )
-from cognite_toolkit._cdf_tk.load import (
+from cognite_toolkit._cdf_tk.loaders import (
     LOADER_BY_FOLDER_NAME,
     NodeLoader,
     TransformationLoader,
