+++ cognite_toolkit/_cdf_tk/commands/pull.py
@@ -21,8 +21,8 @@ from cognite_toolkit._cdf_tk.exceptions import (
     ToolkitMissingResourceError,
     ToolkitNotADirectoryError,
 )
-from cognite_toolkit._cdf_tk.load import ResourceLoader
-from cognite_toolkit._cdf_tk.load._base_loaders import T_ID, T_WritableCogniteResourceList
+from cognite_toolkit._cdf_tk.loaders import ResourceLoader
+from cognite_toolkit._cdf_tk.loaders._base_loaders import T_ID, T_WritableCogniteResourceList
 from cognite_toolkit._cdf_tk.templates.data_classes import BuildConfigYAML, SystemYAML
 from cognite_toolkit._cdf_tk.utils import CDFToolConfig, YAMLComment, YAMLWithComments, tmp_build_directory
