+++ tests/tests_unit/test_cdf_tk/test_describe.py
@@ -6,7 +6,7 @@ from unittest.mock import MagicMock
 from cognite.client.data_classes.data_modeling import Container, DataModel, Space, View
 
 from cognite_toolkit._cdf_tk.commands.describe import describe_datamodel
-from cognite_toolkit._cdf_tk.load import (
+from cognite_toolkit._cdf_tk.loaders import (
     ContainerLoader,
     DataModelLoader,
     SpaceLoader,
