+++ tests/tests_unit/test_cdf_tk/test_parameters/test_functions.py
@@ -14,7 +14,7 @@ from cognite_toolkit._cdf_tk._parameters import (
     read_parameter_from_init_type_hints,
     read_parameters_from_dict,
 )
-from cognite_toolkit._cdf_tk.load import RESOURCE_LOADER_LIST, ResourceLoader
+from cognite_toolkit._cdf_tk.loaders import RESOURCE_LOADER_LIST, ResourceLoader
 
 
 class TestReadParameterFromTypeHints:
