+++ tests/tests_unit/test_cdf_tk/test_templates.py
@@ -8,7 +8,7 @@ import pytest
 import yaml
 
 from cognite_toolkit._cdf_tk.commands.build import BuildCommand, _BuildState, _Helpers
-from cognite_toolkit._cdf_tk.load import LOADER_BY_FOLDER_NAME
+from cognite_toolkit._cdf_tk.loaders import LOADER_BY_FOLDER_NAME
 from cognite_toolkit._cdf_tk.templates import (
     flatten_dict,
     iterate_modules,
