+++ tests/tests_unit/test_cdf_tk/test_validation/test_read_yaml.py
@@ -6,7 +6,7 @@ import yaml
 from cognite.client.data_classes import TimeSeries
 
 from cognite_toolkit._cdf_tk._parameters import ParameterSpecSet
-from cognite_toolkit._cdf_tk.load import ContainerLoader, SpaceLoader, TimeSeriesLoader, ViewLoader
+from cognite_toolkit._cdf_tk.loaders import ContainerLoader, SpaceLoader, TimeSeriesLoader, ViewLoader
 from cognite_toolkit._cdf_tk.tk_warnings import (
     CaseTypoWarning,
     DataSetMissingWarning,
